Move sort_utils drom DisplayTable dir to client root
