+++ client/src/components/DisplayTable/index.js
@@ -4,4 +4,3 @@ export {
   DisplayTableColumnToggle,
   DisplayTableDownloadCsv,
 } from "./DisplayTableUtils";
-export { default_sort_func } from "./sort_utils";
