+++ client/src/panels/panel_declarations/results/result_flat_table.js
@@ -5,12 +5,7 @@ import { HeightClippedGraph } from "src/panels/panel_declarations/common_panel_c
 import { declare_panel } from "src/panels/panel_declarations/common_panel_utils";
 import { InfographicPanel } from "src/panels/panel_declarations/InfographicPanel";
 
-import {
-  LeafSpinner,
-  DisplayTable,
-  default_sort_func,
-  ModalButton,
-} from "src/components/index";
+import { LeafSpinner, DisplayTable, ModalButton } from "src/components/index";
 
 import { businessConstants } from "src/models/businessConstants";
 import * as Results from "src/models/results";
@@ -22,6 +17,7 @@ import { lang } from "src/core/injected_build_constants";
 import { toggle_list } from "src/general_utils";
 import { infograph_href_template } from "src/infographic/infographic_link";
 import { get_source_links } from "src/metadata/data_sources";
+import { default_sort_func } from "src/sort_utils";
 
 import IndicatorDisplayPanel from "./IndicatorDisplayPanel";
